中奖金额小数

+++ src/main/java/com/jgw/supercodeplatform/marketing/dto/activity/MarketingMembersWinRecordListReturn.java
@@ -19,7 +19,7 @@ public class MarketingMembersWinRecordListReturn {
     @ApiModelProperty(value = "中奖奖次id")
     private Long prizeTypeId;//中奖奖次id
     @ApiModelProperty(value = "中奖金额")
-    private Integer winningAmount;//中奖金额
+    private Float winningAmount;//中奖金额
     @ApiModelProperty(value = "中奖码")
     private String winningCode;//中奖码
     @ApiModelProperty(value = "组织id")
@@ -88,11 +88,11 @@ public class MarketingMembersWinRecordListReturn {
         this.prizeTypeId = prizeTypeId;
     }
 
-    public Integer getWinningAmount() {
+    public Float getWinningAmount() {
         return winningAmount;
     }
 
-    public void setWinningAmount(Integer winningAmount) {
+    public void setWinningAmount(Float winningAmount) {
         this.winningAmount = winningAmount;
     }
